componentized page
